LPS-73562 Extract runtime LPKG creation to public util
