+++ modules/apps/static/portal-lpkg-deployer/portal-lpkg-deployer-test-util/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Portal LPKG Deployer Test Utilities
+Bundle-SymbolicName: com.liferay.portal.lpkg.deployer.test.util
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.portal.lpkg.deployer.test.util
