+++ modules/apps/static/portal-lpkg-deployer/portal-lpkg-deployer-test-util/build.gradle
@@ -1,0 +1,8 @@
+dependencies {
+	provided group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.6.0"
+	provided group: "org.osgi", name: "org.osgi.core", version: "5.0.0"
+}
+
+liferay {
+	deployDir = file("${liferayHome}/osgi/test")
+}
