+++ modules/apps/static/portal-lpkg-deployer/portal-lpkg-deployer-test-util/src/main/java/com/liferay/portal/lpkg/deployer/test/util/LPKGTestUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.lpkg.deployer.util;
+package com.liferay.portal.lpkg.deployer.test.util;
 
 import com.liferay.portal.kernel.io.unsync.UnsyncByteArrayInputStream;
 import com.liferay.portal.kernel.io.unsync.UnsyncByteArrayOutputStream;
