+++ modules/apps/static/portal-lpkg-deployer/portal-lpkg-deployer-test/build.gradle
@@ -6,6 +6,7 @@ dependencies {
 	provided project(":apps:static:portal-lpkg-deployer:portal-lpkg-deployer")
 
 	testCompile group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "2.6.0"
+	testCompile project(":apps:static:portal-lpkg-deployer:portal-lpkg-deployer-test-util")
 
 	testIntegrationCompile group: "com.liferay", name: "com.liferay.arquillian.extension.junit.bridge", version: "1.0.7"
 	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.6.0"
